ref(sdk-crashes): Move cocoa tests to extra file
Move SDK crash detection tests related to the Cocoa SDK to extra test file.
